move cjs and esm to root folder
